Move normalize
