+++ modules/normalize.py
@@ -1,5 +1,16 @@
 import torch
 from torch import nn
+from torch.nn import functional as F
+
+
+class L2NormalizationLayer(nn.Module):
+    def __init__(self, dim=-1, eps=1e-12):
+        super().__init__()
+        self.dim = dim
+        self.eps = eps
+
+    def forward(self, x):
+        return F.normalize(x, p=2, dim=self.dim, eps=self.eps)
 
 
 class RMSNorm(nn.Module):
